+++ packages/auth/amplify_auth_cognito_dart/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.11.15
+
+### Features
+- feat(auth): add support for cognito oidc parameters in managed login ([#6442](https://github.com/aws-amplify/amplify-flutter/pull/6442))
+
 ## 0.11.14
 
 ### Features