+++ packages/common/amplify_db_common_dart/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.4.14
+
+- Minor bug fixes and improvements
+
 ## 0.4.13
 
 ### Chores